% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbseqCUDA.R
\name{fbseqCUDA}
\alias{fbseqCUDA}
\title{Function \code{fbseqCUDA}}
\usage{
fbseqCUDA(chain)
}
\arguments{
\item{chain}{object of type \code{Chain}. This \code{chain} argument could be
a newly created \code{Chain} object from \code{Chain(...)}. Alternatively,
if \code{chain} is the output from a previous call to \code{fbseq(...)},
then the function will continue the MCMC from where it left off.}
}
\value{
a \code{Chain} object with updated parameter values. You can feed
this object back into another call to \code{fbseq} to continue the chain
from where you left off.
}
\description{
Calls the C++/CUDA code behind the \code{fbseq()} function
in the \code{fbseq} package.
}

