% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heterosisCUDA.R
\name{heterosisCUDA}
\alias{heterosisCUDA}
\title{Function \code{heterosisCUDA}}
\usage{
heterosisCUDA(chain)
}
\arguments{
\item{chain}{object of type \code{Chain}. This \code{chain} argument could be
a newly created \code{Chain} object from \code{Chain(...)}. Alternatively,
if \code{chain} is the output from a previous call to \code{heterosis(...)},
then the function will continue the MCMC from where it left off.}
}
\value{
a \code{Chain} object with updated parameter values. You can feed
this object back into another call to \code{heterosis} to continue the chain
from where you left off.
}
\description{
Calls the C++/CUDA code behind the \code{run_mcmc()} function
in the \code{heterosis} package.
}

