% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heterosisCUDA.R
\name{RgetDeviceCount}
\alias{RgetDeviceCount}
\title{Function \code{RgetDeviceCount}}
\usage{
RgetDeviceCount()
}
\value{
Number of CUDA-capable GPUs.
}
\description{
Get the number of CUDA-capable GPUs.
}

