% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbseqCUDA.R
\name{RsetDevice}
\alias{RsetDevice}
\title{Function \code{RsetDevice}}
\usage{
RsetDevice(device)
}
\arguments{
\item{device}{Index of the GPU to use. Must be an integer from 0 to number of GPUs - 1.}
}
\value{
Integer index of the current CUDA-capable GPU, which is >= 0 and 
< number of devices.
}
\description{
Set the GPU for running the current MCMC chain.
}

